+++ asm/non_matchings/unknown_0B8920/func_800C01D8.s
@@ -1,3 +1,8 @@
+.late_rodata
+glabel jpt_800E92D0
+.word L800C0310, L800C0320, L800C0370, L800C03C0, L800C03D0, L800C0420, L800C0470
+
+.text
 glabel func_800C01D8
 /* 0C0DD8 800C01D8 3C0E800E */  lui   $t6, %hi(D_800E31A0) # $t6, 0x800e
 /* 0C0DDC 800C01DC 8DCE31A0 */  lw    $t6, %lo(D_800E31A0)($t6)
@@ -74,16 +79,18 @@ glabel func_800C01D8
 /* 0C0EEC 800C02EC 10200062 */  beqz  $at, .L800C0478
 /* 0C0EF0 800C02F0 A0AD0000 */   sb    $t5, ($a1)
 /* 0C0EF4 800C02F4 000E7080 */  sll   $t6, $t6, 2
-/* 0C0EF8 800C02F8 3C01800F */  lui   $at, %hi(D_800E92D0) # $at, 0x800f
+/* 0C0EF8 800C02F8 3C01800F */  lui   $at, %hi(jpt_800E92D0) # $at, 0x800f
 /* 0C0EFC 800C02FC 002E0821 */  addu  $at, $at, $t6
-/* 0C0F00 800C0300 8C2E92D0 */  lw    $t6, %lo(D_800E92D0)($at)
+/* 0C0F00 800C0300 8C2E92D0 */  lw    $t6, %lo(jpt_800E92D0)($at)
 /* 0C0F04 800C0304 00000000 */  nop   
 /* 0C0F08 800C0308 01C00008 */  jr    $t6
 /* 0C0F0C 800C030C 00000000 */   nop   
+glabel L800C0310
 /* 0C0F10 800C0310 0C0301E0 */  jal   func_800C0780
 /* 0C0F14 800C0314 00000000 */   nop   
 /* 0C0F18 800C0318 10000057 */  b     .L800C0478
 /* 0C0F1C 800C031C 00000000 */   nop   
+glabel L800C0320
 /* 0C0F20 800C0320 3C02800E */  lui   $v0, %hi(D_800E32D0) # $v0, 0x800e
 /* 0C0F24 800C0324 3C0F800E */  lui   $t7, %hi(D_800E32A0) # $t7, 0x800e
 /* 0C0F28 800C0328 3C18800E */  lui   $t8, %hi(D_800E32AC) # $t8, 0x800e
@@ -104,6 +111,7 @@ glabel func_800C01D8
 /* 0C0F64 800C0364 24060008 */   li    $a2, 8
 /* 0C0F68 800C0368 10000043 */  b     .L800C0478
 /* 0C0F6C 800C036C 00000000 */   nop   
+glabel L800C0370
 /* 0C0F70 800C0370 3C02800E */  lui   $v0, %hi(D_800E32D0) # $v0, 0x800e
 /* 0C0F74 800C0374 3C08800E */  lui   $t0, %hi(D_800E32A0) # $t0, 0x800e
 /* 0C0F78 800C0378 3C09800E */  lui   $t1, %hi(D_800E32AC) # $t1, 0x800e
@@ -124,10 +132,12 @@ glabel func_800C01D8
 /* 0C0FB4 800C03B4 24060008 */   li    $a2, 8
 /* 0C0FB8 800C03B8 1000002F */  b     .L800C0478
 /* 0C0FBC 800C03BC 00000000 */   nop   
+glabel L800C03C0
 /* 0C0FC0 800C03C0 0C030575 */  jal   func_800C15D4
 /* 0C0FC4 800C03C4 00000000 */   nop   
 /* 0C0FC8 800C03C8 1000002B */  b     .L800C0478
 /* 0C0FCC 800C03CC 00000000 */   nop   
+glabel L800C03D0
 /* 0C0FD0 800C03D0 3C02800E */  lui   $v0, %hi(D_800E34F8) # $v0, 0x800e
 /* 0C0FD4 800C03D4 3C0B800E */  lui   $t3, %hi(D_800E349C) # $t3, 0x800e
 /* 0C0FD8 800C03D8 3C0C800E */  lui   $t4, %hi(D_800E3440) # $t4, 0x800e
@@ -148,6 +158,7 @@ glabel func_800C01D8
 /* 0C1014 800C0414 24060050 */   li    $a2, 80
 /* 0C1018 800C0418 10000017 */  b     .L800C0478
 /* 0C101C 800C041C 00000000 */   nop   
+glabel L800C0420
 /* 0C1020 800C0420 3C02800E */  lui   $v0, %hi(D_800E3338) # $v0, 0x800e
 /* 0C1024 800C0424 3C0E800E */  lui   $t6, %hi(D_800E330C) # $t6, 0x800e
 /* 0C1028 800C0428 3C0F800E */  lui   $t7, %hi(D_800E3318) # $t7, 0x800e
@@ -168,6 +179,7 @@ glabel func_800C01D8
 /* 0C1064 800C0464 24060006 */   li    $a2, 6
 /* 0C1068 800C0468 10000003 */  b     .L800C0478
 /* 0C106C 800C046C 00000000 */   nop   
+glabel L800C0470
 /* 0C1070 800C0470 0C030990 */  jal   func_800C2640
 /* 0C1074 800C0474 00000000 */   nop   
 .L800C0478:
