+++ asm/non_matchings/unknown_0B8920/func_800C0494.s
@@ -1,3 +1,8 @@
+.late_rodata
+glabel jpt_800E92EC
+.word L800C056C, L800C0580, L800C0580, L800C0594, L800C0580, L800C0580, L800C05A8
+
+.text
 glabel func_800C0494
 /* 0C1094 800C0494 3C03800E */  lui   $v1, %hi(D_800E31A4) # $v1, 0x800e
 /* 0C1098 800C0498 246331A4 */  addiu $v1, %lo(D_800E31A4) # addiu $v1, $v1, 0x31a4
@@ -52,27 +57,31 @@ glabel func_800C0494
 /* 0C1148 800C0548 2D610007 */  sltiu $at, $t3, 7
 /* 0C114C 800C054C 1020001A */  beqz  $at, .L800C05B8
 /* 0C1150 800C0550 000B5880 */   sll   $t3, $t3, 2
-/* 0C1154 800C0554 3C01800F */  lui   $at, %hi(D_800E92EC) # $at, 0x800f
+/* 0C1154 800C0554 3C01800F */  lui   $at, %hi(jpt_800E92EC) # $at, 0x800f
 /* 0C1158 800C0558 002B0821 */  addu  $at, $at, $t3
-/* 0C115C 800C055C 8C2B92EC */  lw    $t3, %lo(D_800E92EC)($at)
+/* 0C115C 800C055C 8C2B92EC */  lw    $t3, %lo(jpt_800E92EC)($at)
 /* 0C1160 800C0560 00000000 */  nop   
 /* 0C1164 800C0564 01600008 */  jr    $t3
 /* 0C1168 800C0568 00000000 */   nop   
+glabel L800C056C
 /* 0C116C 800C056C 0C03020D */  jal   func_800C0834
 /* 0C1170 800C0570 00000000 */   nop   
 /* 0C1174 800C0574 3C03800E */  lui   $v1, %hi(D_800E31AC) # $v1, 0x800e
 /* 0C1178 800C0578 1000000F */  b     .L800C05B8
 /* 0C117C 800C057C 246331AC */   addiu $v1, %lo(D_800E31AC) # addiu $v1, $v1, 0x31ac
+glabel L800C0580
 /* 0C1180 800C0580 0C03044C */  jal   func_800C1130
 /* 0C1184 800C0584 00000000 */   nop   
 /* 0C1188 800C0588 3C03800E */  lui   $v1, %hi(D_800E31AC) # $v1, 0x800e
 /* 0C118C 800C058C 1000000A */  b     .L800C05B8
 /* 0C1190 800C0590 246331AC */   addiu $v1, %lo(D_800E31AC) # addiu $v1, $v1, 0x31ac
+glabel L800C0594
 /* 0C1194 800C0594 0C0307BA */  jal   func_800C1EE8
 /* 0C1198 800C0598 00000000 */   nop   
 /* 0C119C 800C059C 3C03800E */  lui   $v1, %hi(D_800E31AC) # $v1, 0x800e
 /* 0C11A0 800C05A0 10000005 */  b     .L800C05B8
 /* 0C11A4 800C05A4 246331AC */   addiu $v1, %lo(D_800E31AC) # addiu $v1, $v1, 0x31ac
+glabel L800C05A8
 /* 0C11A8 800C05A8 0C0309E8 */  jal   func_800C27A0
 /* 0C11AC 800C05AC 00000000 */   nop   
 /* 0C11B0 800C05B0 3C03800E */  lui   $v1, %hi(D_800E31AC) # $v1, 0x800e
