+++ asm/non_matchings/unknown_0B8920/func_800C05C8.s
@@ -1,3 +1,12 @@
+.late_rodata
+glabel D_800E9308
+.float 1.4
+glabel D_800E930C
+.float 1.2
+glabel jpt_800E9310
+.word L800C0680, L800C0694, L800C06A8, L800C06BC, L800C06D0, L800C06E4, L800C06F8
+
+.text
 glabel func_800C05C8
 /* 0C11C8 800C05C8 27BDFFE0 */  addiu $sp, $sp, -0x20
 /* 0C11CC 800C05CC 3C0E800E */  lui   $t6, %hi(D_800E31AC) # $t6, 0x800e
@@ -41,42 +50,49 @@ glabel func_800C05C8
 /* 0C125C 800C065C 2F010007 */  sltiu $at, $t8, 7
 /* 0C1260 800C0660 10200028 */  beqz  $at, .L800C0704
 /* 0C1264 800C0664 0018C080 */   sll   $t8, $t8, 2
-/* 0C1268 800C0668 3C01800F */  lui   $at, %hi(D_800E9310) # $at, 0x800f
+/* 0C1268 800C0668 3C01800F */  lui   $at, %hi(jpt_800E9310) # $at, 0x800f
 /* 0C126C 800C066C 00380821 */  addu  $at, $at, $t8
-/* 0C1270 800C0670 8C389310 */  lw    $t8, %lo(D_800E9310)($at)
+/* 0C1270 800C0670 8C389310 */  lw    $t8, %lo(jpt_800E9310)($at)
 /* 0C1274 800C0674 00000000 */  nop   
 /* 0C1278 800C0678 03000008 */  jr    $t8
 /* 0C127C 800C067C 00000000 */   nop   
+glabel L800C0680
 /* 0C1280 800C0680 02002025 */  move  $a0, $s0
 /* 0C1284 800C0684 0C030282 */  jal   func_800C0A08
 /* 0C1288 800C0688 02202825 */   move  $a1, $s1
 /* 0C128C 800C068C 1000001E */  b     .L800C0708
 /* 0C1290 800C0690 02002025 */   move  $a0, $s0
+glabel L800C0694
 /* 0C1294 800C0694 02002025 */  move  $a0, $s0
 /* 0C1298 800C0698 0C0304F9 */  jal   func_800C13E4
 /* 0C129C 800C069C 02202825 */   move  $a1, $s1
 /* 0C12A0 800C06A0 10000019 */  b     .L800C0708
 /* 0C12A4 800C06A4 02002025 */   move  $a0, $s0
+glabel L800C06A8
 /* 0C12A8 800C06A8 02002025 */  move  $a0, $s0
 /* 0C12AC 800C06AC 0C030537 */  jal   func_800C14DC
 /* 0C12B0 800C06B0 02202825 */   move  $a1, $s1
 /* 0C12B4 800C06B4 10000014 */  b     .L800C0708
 /* 0C12B8 800C06B8 02002025 */   move  $a0, $s0
+glabel L800C06BC
 /* 0C12BC 800C06BC 02002025 */  move  $a0, $s0
 /* 0C12C0 800C06C0 0C03089D */  jal   func_800C2274
 /* 0C12C4 800C06C4 02202825 */   move  $a1, $s1
 /* 0C12C8 800C06C8 1000000F */  b     .L800C0708
 /* 0C12CC 800C06CC 02002025 */   move  $a0, $s0
+glabel L800C06D0
 /* 0C12D0 800C06D0 02002025 */  move  $a0, $s0
 /* 0C12D4 800C06D4 0C0308FE */  jal   func_800C23F8
 /* 0C12D8 800C06D8 02202825 */   move  $a1, $s1
 /* 0C12DC 800C06DC 1000000A */  b     .L800C0708
 /* 0C12E0 800C06E0 02002025 */   move  $a0, $s0
+glabel L800C06E4
 /* 0C12E4 800C06E4 02002025 */  move  $a0, $s0
 /* 0C12E8 800C06E8 0C030952 */  jal   func_800C2548
 /* 0C12EC 800C06EC 02202825 */   move  $a1, $s1
 /* 0C12F0 800C06F0 10000005 */  b     .L800C0708
 /* 0C12F4 800C06F4 02002025 */   move  $a0, $s0
+glabel L800C06F8
 /* 0C12F8 800C06F8 02002025 */  move  $a0, $s0
 /* 0C12FC 800C06FC 0C030A3A */  jal   func_800C28E8
 /* 0C1300 800C0700 02202825 */   move  $a1, $s1
