+++ asm/non_matchings/unknown_0B8920/func_800C38B4.s
@@ -1,3 +1,8 @@
+.late_rodata
+glabel jpt_800E9330
+.word L800C3958, L800C3984, L800C3A30, L800C3AA8, L800C3ADC, L800C3B00, L800C3B2C, L800C3B50, L800C3B88, L800C3BA0
+
+.text
 glabel func_800C38B4
 /* 0C44B4 800C38B4 27BDFFB8 */  addiu $sp, $sp, -0x48
 /* 0C44B8 800C38B8 AFB3002C */  sw    $s3, 0x2c($sp)
@@ -35,12 +40,13 @@ glabel func_800C38B4
 /* 0C4534 800C3934 2DE1000A */  sltiu $at, $t7, 0xa
 /* 0C4538 800C3938 1020009E */  beqz  $at, .L800C3BB4
 /* 0C453C 800C393C 000F7880 */   sll   $t7, $t7, 2
-/* 0C4540 800C3940 3C01800F */  lui   $at, %hi(D_800E9330) # $at, 0x800f
+/* 0C4540 800C3940 3C01800F */  lui   $at, %hi(jpt_800E9330) # $at, 0x800f
 /* 0C4544 800C3944 002F0821 */  addu  $at, $at, $t7
-/* 0C4548 800C3948 8C2F9330 */  lw    $t7, %lo(D_800E9330)($at)
+/* 0C4548 800C3948 8C2F9330 */  lw    $t7, %lo(jpt_800E9330)($at)
 /* 0C454C 800C394C 00000000 */  nop   
 /* 0C4550 800C3950 01E00008 */  jr    $t7
 /* 0C4554 800C3954 00000000 */   nop   
+glabel L800C3958
 /* 0C4558 800C3958 92050001 */  lbu   $a1, 1($s0)
 /* 0C455C 800C395C 26310002 */  addiu $s1, $s1, 2
 /* 0C4560 800C3960 24040001 */  li    $a0, 1
@@ -52,6 +58,7 @@ glabel func_800C38B4
 /* 0C4578 800C3978 92020000 */  lbu   $v0, ($s0)
 /* 0C457C 800C397C 1000008E */  b     .L800C3BB8
 /* 0C4580 800C3980 28410003 */   slti  $at, $v0, 3
+glabel L800C3984
 /* 0C4584 800C3984 92090001 */  lbu   $t1, 1($s0)
 /* 0C4588 800C3988 3C0E8000 */  lui   $t6, %hi(osTvType) # $t6, 0x8000
 /* 0C458C 800C398C AE490004 */  sw    $t1, 4($s2)
@@ -97,6 +104,7 @@ glabel func_800C38B4
 /* 0C4624 800C3A24 92020000 */  lbu   $v0, ($s0)
 /* 0C4628 800C3A28 10000063 */  b     .L800C3BB8
 /* 0C462C 800C3A2C 28410003 */   slti  $at, $v0, 3
+glabel L800C3A30
 /* 0C4630 800C3A30 92050001 */  lbu   $a1, 1($s0)
 /* 0C4634 800C3A34 24040001 */  li    $a0, 1
 /* 0C4638 800C3A38 AE450014 */  sw    $a1, 0x14($s2)
@@ -127,6 +135,7 @@ glabel func_800C38B4
 /* 0C469C 800C3A9C 92020000 */  lbu   $v0, ($s0)
 /* 0C46A0 800C3AA0 10000045 */  b     .L800C3BB8
 /* 0C46A4 800C3AA4 28410003 */   slti  $at, $v0, 3
+glabel L800C3AA8
 /* 0C46A8 800C3AA8 920C0001 */  lbu   $t4, 1($s0)
 /* 0C46AC 800C3AAC 00000000 */  nop   
 /* 0C46B0 800C3AB0 15800003 */  bnez  $t4, .L800C3AC0
@@ -142,6 +151,7 @@ glabel func_800C38B4
 /* 0C46D0 800C3AD0 92020000 */  lbu   $v0, ($s0)
 /* 0C46D4 800C3AD4 10000038 */  b     .L800C3BB8
 /* 0C46D8 800C3AD8 28410003 */   slti  $at, $v0, 3
+glabel L800C3ADC
 /* 0C46DC 800C3ADC 920E0001 */  lbu   $t6, 1($s0)
 /* 0C46E0 800C3AE0 26310002 */  addiu $s1, $s1, 2
 /* 0C46E4 800C3AE4 AE4E0028 */  sw    $t6, 0x28($s2)
@@ -151,6 +161,7 @@ glabel func_800C38B4
 /* 0C46F4 800C3AF4 92020000 */  lbu   $v0, ($s0)
 /* 0C46F8 800C3AF8 1000002F */  b     .L800C3BB8
 /* 0C46FC 800C3AFC 28410003 */   slti  $at, $v0, 3
+glabel L800C3B00
 /* 0C4700 800C3B00 8E58002C */  lw    $t8, 0x2c($s2)
 /* 0C4704 800C3B04 92190001 */  lbu   $t9, 1($s0)
 /* 0C4708 800C3B08 26310002 */  addiu $s1, $s1, 2
@@ -162,6 +173,7 @@ glabel func_800C38B4
 /* 0C4720 800C3B20 92020000 */  lbu   $v0, ($s0)
 /* 0C4724 800C3B24 10000024 */  b     .L800C3BB8
 /* 0C4728 800C3B28 28410003 */   slti  $at, $v0, 3
+glabel L800C3B2C
 /* 0C472C 800C3B2C 920A0001 */  lbu   $t2, 1($s0)
 /* 0C4730 800C3B30 26310002 */  addiu $s1, $s1, 2
 /* 0C4734 800C3B34 AE4A0030 */  sw    $t2, 0x30($s2)
@@ -171,6 +183,7 @@ glabel func_800C38B4
 /* 0C4744 800C3B44 92020000 */  lbu   $v0, ($s0)
 /* 0C4748 800C3B48 1000001B */  b     .L800C3BB8
 /* 0C474C 800C3B4C 28410003 */   slti  $at, $v0, 3
+glabel L800C3B50
 /* 0C4750 800C3B50 828C0000 */  lb    $t4, ($s4)
 /* 0C4754 800C3B54 2404003C */  li    $a0, 60
 /* 0C4758 800C3B58 15800007 */  bnez  $t4, .L800C3B78
@@ -186,12 +199,14 @@ glabel func_800C38B4
 /* 0C477C 800C3B7C 26310002 */  addiu $s1, $s1, 2
 /* 0C4780 800C3B80 1000000C */  b     .L800C3BB4
 /* 0C4784 800C3B84 26100002 */   addiu $s0, $s0, 2
+glabel L800C3B88
 /* 0C4788 800C3B88 92180001 */  lbu   $t8, 1($s0)
 /* 0C478C 800C3B8C 26310002 */  addiu $s1, $s1, 2
 /* 0C4790 800C3B90 A2D80000 */  sb    $t8, ($s6)
 /* 0C4794 800C3B94 92020002 */  lbu   $v0, 2($s0)
 /* 0C4798 800C3B98 10000006 */  b     .L800C3BB4
 /* 0C479C 800C3B9C 26100002 */   addiu $s0, $s0, 2
+glabel L800C3BA0
 /* 0C47A0 800C3BA0 92190001 */  lbu   $t9, 1($s0)
 /* 0C47A4 800C3BA4 26310002 */  addiu $s1, $s1, 2
 /* 0C47A8 800C3BA8 A2B90000 */  sb    $t9, ($s5)
